# fern/definition/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Airweave",
    "version": "0.1.0",
    "description": "\n\n## API Groups\nThis API spec only includes the following API groups:\n- **Sources**: API endpoints for discovering available data source connectors and their configuration requirements\n- **Collections**: API endpoints for managing collections - logical groups of data sources that provide unified search capabilities\n- **Source Connections**: API endpoints for managing live connections to data sources. Source connections are the actual configured instances that Airweave uses to sync data from your apps and databases, transforming it into searchable, structured information within collections\n- **Events**: API endpoints for managing webhook subscriptions and event messages. Subscribe to events like sync completions to receive real-time notifications at your webhook URL\n"
  },
  "paths": {
...
    "/events/messages": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "List Messages",
        "description": "Retrieve all event messages for your organization.\n\nEvent messages represent webhook payloads that were sent (or attempted to be sent)\nto your subscribed endpoints. Each message contains the event type, payload data,\nand delivery status information.\n\nUse the `event_types` query parameter to filter messages by specific event types,\nsuch as `sync.completed` or `sync.failed`.",
        "operationId": "get_messages_events_messages_get",
        "parameters": [
          {
            "name": "event_types",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter messages by event type(s). Accepts multiple values, e.g., `?event_types=sync.completed&event_types=sync.failed`.",
              "example": [
                "sync.completed",
                "sync.failed"
              ],
              "title": "Event Types"
            },
            "description": "Filter messages by event type(s). Accepts multiple values, e.g., `?event_types=sync.completed&event_types=sync.failed`."
          }
        ],
        "responses": {
          "200": {
            "description": "List of event messages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventMessage"
                  },
                  "title": "Response 200 Get Messages Events Messages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate Limit Exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/events/messages/{message_id}": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get Message",
        "description": "Retrieve a specific event message by its ID.\n\nReturns the full message details including the event type, payload data,\ntimestamp, and delivery channel information. Use this to inspect the\nexact payload that was sent to your webhook endpoints.\n\nUse `include_attempts=true` to also retrieve delivery attempts for this message,\nwhich include HTTP response codes, response bodies, and timestamps for debugging\ndelivery failures.",
        "operationId": "get_message_events_messages__message_id__get",
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The unique identifier of the message to retrieve (UUID).",
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "title": "Message Id"
            },
            "description": "The unique identifier of the message to retrieve (UUID)."
          },
          {
            "name": "include_attempts",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include delivery attempts for this message. Each attempt includes the HTTP response code, response body, and timestamp.",
              "default": false,
              "title": "Include Attempts"
            },
            "description": "Include delivery attempts for this message. Each attempt includes the HTTP response code, response body, and timestamp."
          }
        ],
        "responses": {
          "200": {
            "description": "Event message details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventMessage"
                }
              }
            }
          },
          "404": {
            "description": "Message Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate Limit Exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/events/subscriptions": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "List Subscriptions",
        "description": "List all webhook subscriptions for your organization.\n\nReturns all configured webhook endpoints, including their URLs, subscribed\nevent types, and current status (enabled/disabled). Use this to audit\nyour webhook configuration or find a specific subscription.",
        "operationId": "get_subscriptions_events_subscriptions_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of webhook subscriptions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WebhookSubscription"
                  },
                  "title": "Response 200 Get Subscriptions Events Subscriptions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate Limit Exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "events"
        ],
        "summary": "Create Subscription",
        "description": "Create a new webhook subscription.\n\nWebhook subscriptions allow you to receive real-time notifications when events\noccur in Airweave. When you create a subscription, you specify:\n\n- **URL**: The HTTPS endpoint where events will be delivered\n- **Event Types**: Which events you want to receive (e.g., `sync.completed`, `sync.failed`)\n- **Secret** (optional): A custom signing secret for verifying webhook signatures\n\nAfter creation, Airweave will send HTTP POST requests to your URL whenever\nmatching events occur. Each request includes a signature header for verification.",
        "operationId": "create_subscription_events_subscriptions_post",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubscriptionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created subscription",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookSubscription"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate Limit Exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/events/subscriptions/{subscription_id}/recover": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "Recover Failed Messages",
        "description": "Retry failed message deliveries for a webhook subscription.\n\nTriggers a recovery process that replays all failed messages within the\nspecified time window. This is useful when:\n\n- Your endpoint was temporarily down and you want to catch up\n- You've fixed a bug in your webhook handler\n- You want to reprocess events after re-enabling a disabled subscription\n\nMessages are retried in chronological order. Successfully delivered messages\nare skipped; only failed or pending messages are retried.",
        "operationId": "recover_failed_messages_events_subscriptions__subscription_id__recover_post",
        "parameters": [
          {
            "name": "subscription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The unique identifier of the subscription to recover messages for (UUID).",
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "title": "Subscription Id"
            },
            "description": "The unique identifier of the subscription to recover messages for (UUID)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoverMessagesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Recovery task information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryTask"
                }
              }
            }
          },
          "404": {
            "description": "Subscription Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate Limit Exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "EventMessage": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique identifier for this message (UUID format)",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type",
            "description": "The type of event (e.g., 'sync.completed', 'sync.failed')",
            "example": "sync.completed"
          },
          "payload": {
            "$ref": "#/components/schemas/SyncEventPayload",
            "description": "The event payload data, matching what is delivered to webhooks"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "When this message was created (ISO 8601 format, UTC)",
            "example": "2024-03-15T09:45:32Z"
          },
          "channels": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channels",
            "description": "Channels this message was sent to (typically matches the event type)",
            "example": [
              "sync.completed"
            ]
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags",
            "description": "Tags associated with this message for filtering"
          }
        },
        "type": "object",
        "required": [
          "id",
          "event_type",
          "payload",
          "timestamp"
        ],
        "title": "EventMessage",
        "description": "An event message that was sent (or attempted) to webhook subscribers.\n\nThe payload contains the actual event data matching the webhook delivery format.",
        "example": {
          "channels": [
            "sync.completed"
          ],
          "event_type": "sync.completed",
          "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
          "payload": {
            "collection_name": "Customer Support Tickets",
            "collection_readable_id": "customer-support-tickets-x7k9m",
            "event_type": "sync.completed",
            "job_id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
            "source_connection_id": "b2c3d4e5-f6a7-8901-bcde-f23456789012",
            "source_type": "zendesk",
            "status": "completed",
            "timestamp": "2024-03-15T09:45:32Z"
          },
          "timestamp": "2024-03-15T09:45:32Z"
        }
      },
...
      "NotFoundErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Error message describing what was not found",
            "example": "Subscription not found"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "NotFoundErrorResponse",
        "description": "Response returned when a resource is not found (HTTP 404).",
        "example": {
          "detail": "Subscription not found"
        }
      },
...
      "RateLimitErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Error message explaining the rate limit",
            "example": "Rate limit exceeded. Please retry after 60 seconds."
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "RateLimitErrorResponse",
        "description": "Response returned when rate limit is exceeded (HTTP 429).\n\nThe API enforces rate limits to ensure fair usage. When exceeded,\nwait for the duration specified in the Retry-After header before retrying.",
        "example": {
          "detail": "Rate limit exceeded. Please retry after 60 seconds."
        }
      },
...
      "ValidationErrorResponse": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationErrorDetail"
            },
            "type": "array",
            "title": "Detail",
            "description": "List of validation errors"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ValidationErrorResponse",
        "description": "Response returned when request validation fails (HTTP 422).\n\nThis occurs when the request body contains invalid data, such as\nmalformed URLs, invalid event types, or missing required fields.",
        "example": {
          "detail": [
            {
              "loc": [
                "body",
                "url"
              ],
              "msg": "Invalid URL: scheme must be http or https",
              "type": "value_error.url.scheme"
            },
            {
              "loc": [
                "body",
                "event_types"
              ],
              "msg": "event_types cannot be empty",
              "type": "value_error"
            }
          ]
        }
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for authentication"
      }
    }
  },
...
  "servers": [
    {
      "url": "https://api.airweave.ai",
      "description": "Production",
      "x-fern-server-name": "Production"
    },
    {
      "url": "http://localhost:8001",
      "description": "Local",
      "x-fern-server-name": "Local"
    }
  ]
}